lmapd: make xml backend optional for production
Allow for compile-time selection of whether to build support for JSON,
XML, or both, and add CLI options to select at runtime which IO backend
to use in lmapd.

For the unit tests to be built, the XML backend must be present.  We may
fix that in the future.

CEPTRO.br-issue: MDC-438