lmapd: support compile-time hardcoding of multiple config
Extend the code to support hardcoding several config files or
directories.  It adds a very small fixed cost to setups with a single
config file or directory.

It is extremely non-trivial to implement this tied to the build
system as a list of strings, so use the well-kown "colon-delimited"
list of paths that is used in POSIX for the shell PATH environment
variable.  We also accept files, not just directories, though.

CEPTRO.br-issue: MDC-452