lmapd: json: do not render nodes that are empty arrays
For the LMAP JSON YANG model, all instances of { "foo":[ ] } are exactly
the same as { }, and our parser code handles it like that as well.

Render these named empty arrays accordingly, suppresing node "foo" when
it is an empty array.  Note that an array with empty elements is *not*
empty, and therefore it will be rendered, e.g. { "foo":[null] }.

This mainly affects reports, e.g. suppressing empty result tables and
"value" arrays inside an empty row (the row now becomes an empty
object), etc.  It also uppresses empty "function" arrays in registries
when rendering the lmapd state/config.

Also, update the unit tests accordingly.

This is a cosmetic fix, but it does reduce the output size and improves
its readability.

CEPTRO.br-issue: MDC-438