lmapd: move action results as a group when the schedule finishes
We want to group in the same report the results of all actions of a
schedule run.  The simplest way to do this is to move the results of an
action to their destinations only when the schedule finishes running,
instead of as soon as the action finishes running.

We need to special-case actions whose destinations are their own
schedule and move the results immediately, so that (on sequential
schedules), they can still be used by the next action in the sequence.

CEPTRO.br-issue: MDC-952