lmapd: do not abort due to too large a random spread
Should the random spread mechanism cause two firings of an event to
accumulate, we'd call event_gaga() on an event that already has
event->fire_event set and active.

This either causes an assertion failure on event_gaga() when trying to
add the second event while the first is still active, or if assertions
are disabled (production build), it overwrites event->fire_event,
forgetting about the already active event.

This causes two "fire" events to be active, something the current code
is not prepared to support.  The obvious issue is that fire_cb() will
run twice for the same (lmapd) struct event, which would crash before
the change of that codepath to safe_event_free().  But there could be
other issues too, such as incorrect event timestamp metadata.

Explicitly detect that an event is already active in event_gaga() and
output an error message (and cope gracefully, by losing the new event).
This ensures that the event that does trigger will be handled correctly
and have correct metadata, etc.