lmapd: refuse too large random spreads for periodic events
Refuse configurations that would knowingly cause periodic event
reordering due to a too large random spread.

It would be nice to detect this also for the calendar event type, but
that's non-trivial, and not worth the hassle now that lmapd has been
changed to handle this gracefully (instead of triggering a crash or
assertion failure).

CEPTRO.br-issue: MDC-1016