lmapd: protect against event_free(NULL)
Internal errors (bugs) could cause lmapd to call event_free() on a NULL
pointer.  event_free() is *not* documented to handle NULL pointers,
unlike free(), and in fact it does not handle them well at all (at least
on libevent2 2.1).

That could cause a NULL pointer derreference, and a crash.  It is better
to cope with it: log the problem, and keep the daemon alive.

This issue has been triggered in production, it can happen.

CEPTRO.br-issue: MDC-1016