lmapctl: report: do not add incomplete/invalid report data
When generating a lmap report, ensure that only task output (output of
an lmap schedule's action) that has been fully loaded without parsing
and validation errors is accepted.

Empty output should be accepted and added to the report, as long as it
is in the proper format for "empty output", and its metadata is correct.

"lmapctl report" will only return an error status if it hits a general
problem (bad config data for example): it will return a success exit
status regardless of how much invalid input data it skips.

Fixes: commit d99c5541196323f1e8b9e9b412022c2dd1d6115b
       (lmapctl: report: prepare to accept structured output from tasks)

CEPTRO.br-issue: MDC-1017
CEPTRO.br-issue: MDC-449