lmapd: runner: run actions on their own process groups
Use setgpid() to isolate each action on their own process group, and
send the SIGTERM signal to that entire process group when the action
needs to be stopped.

This requires a change to use waitpid(-1, ...) instead of waitpid(0,
...) otherwise we will accumulate zombie processes.

Do the setpgid() both on the child and the parent to close the fork()
race as best-practice dictates, but action_kill() as currently
implemented would work correctly either way: should we fail to send
SIGTERM to the process group, we do a second attempt targetting the
action's main process, only.

CEPTRO.br-issue: MDC-1141