lmapd, lmapctl: cli: more flexible config path option
Use an config path argument of "+" to stand for the default built-in
config path.  This allows for easy prepending or appending of config
path components, while still alowing the use of the built-in config
path.

This is useful when one needs more than one instance of lmapd running.

Typical usage would be: "lmapd -c + -c this-instance-config/" to load
the configuration in the built-in path common to all instances of lmapd,
and then load the extra configuration directory after the built-in
configuration, specific to that instance of lmapd.

CEPTRO.br-issue: MDC-1420