deb: change built-in config path, breaks simet-ma before 0.18.7
Remove the last component (the schedule file) from the built-in config
path of lmapd and lmapctl.  simet-ma 0.18.7 will use the new semantics
of the -c option in simet-lmapd 0.15.3 to append the schedule file to
the built-in path, in order to support multiple lmapd instances running
at the same time with different schedules.

This change in the built-in config path is incompatible with simet-ma
before 0.18.7, so add a Breaks header to debian/control.

CEPTRO.br-issue: MDC-1419, MDC-1420