lmapd: skip events of type startup on SIGHUP reload (v2)
Do not run events with a trigger type of 'startup' more than once, even
if a new (or the same) schedule is reloaded through SIGHUP.  This is
closer to the RFC8194 behavior (a lmapd config reload is *not* a
restart, even if internally it is handled as such).

Before this change, lmapd would treat trigger events of type 'startup'
and 'immediate' as effectively the same thing.  They would be triggered
right after starting, and also (if they were still present in the
schedule configuration) after every SIGHUP (i.e. on configuration
reload).

With this change, lmapd will trigger 'startup' events at most once per
session (a session begins when lmapd is started, and ends when it stops
/ terminates).  Once lmapd has triggered at least one LMAP action with
an event type of 'startup', it will disable that trigger *type* for any
*future* configuration reloads / SIGHUP.

Note that it *will* process all 'startup' triggers on the current
configuration, unless interrupted (by a reload, etc).

A lmapd *service* restart (which stops the lmapd process entirely, and
starts it anew) will clear state, and run 'startup' events again if
present in the schedule.

Events of type 'immediate' are unmodified, and will run on start, and
also after configuration reloads.

CEPTRO.br-issue: MDC-1466