source: remove or audit every integer size conversion in data.c
gcc is not smart enough to avoid false-positive warnings about a value
that was previously verified by the code to be in-range for a narrower
type when implicit converting to that type.

Either change variable types to avoid the implicit conversions in the
first place, or ensure the variable with the wider type was verified to
be in-range for the narrower type and add a explicitly typecast to
silence the warning.

CEPTRO.br-issue: MDC-1521