lmapd: add just-in-case null guards to pidfile, atexit() callback
Add asserts to pidfile.c functions, and ensure the lmapd atexit()
callback function won't call the pidfile functions with a NULL lmapd
struct pointer.

This is not fixing any bugs: if the current lmapd main() code fails to
allocate or initialize the lmapd struct, it will terminate the program
*before* it registers the atexit callback or calls any of the pidfile
functions.  It is being done to eliminate a pitfall for future
maintenance.

CEPTRO.br-issue: MDC-1521