xml: remove shadowing of a local variable
Avoid a warning about the harmless shadowing of the local "tab" variable
by renaming the second declaration of the "tab" variable to "table".

The scope where "tab" was shadowed is really small (two lines), and the
code is obviously correct, but the warning needs to go away.

CEPTRO.br-issue: MDC-1521