lmapd: do not loop for too long closing fds after fork
_SC_OPEN_MAX could easily be in the millions range when an ulimit isn't
in place, and doing a million close syscalls is very much *not* cheap.
lmapd should not just loop over all possible FDs calling close() on
them: it is asking for trouble.

closefrom() in modern glibc it will either use the close_range kernel
syscall, or iterate /proc/<pid>/fd/ as a fallback, and abort the program
if neither one works.  BSDs often have closefrom() in libc, and at least
FreeBSD 8 has a kernel syscall for it, with similar semantics.

If closefrom() is not avalable, the fallback close() loop is used.

* CMakeLists: attempt to enable _GNU_SOURCE for closefrom() in glibc
* CMakeLists: detect closefrom() in unistd.h and set HAVE_CLOSEFROM
* lmapd: use closefrom(3) if HAVE_CLOSEFROM is defined
* lmapd: limit the non-closefrom() loop to close only the first 4096 FDs

CEPTRO.br-issue: MDC-1522