csv: ENOMEM errorpath fixes
If realloc() fails due to ENOMEM, csv_next() would derrefrence a NULL
pointer.  It would also leak memory if we just added a braindead error
path, due to the foo = realloc(foo, size) antipattern.

Change xrealloc() to be actually useful and reliabily set errno to
indicate failure.  Unlike realloc(), it will return the original pointer
in that case and thus foo = xrealloc(foo, size) is no longer an
antipattern.

Make csv_next() return NULL on ENOMEM, with errno set, when xrealloc
fails.  Set errno to zero on success.

Get rid of the redundant NULL checking in xfree(): anything that
requires that is not even remotely supported in the first place.

CEPTRO.br-issue: MDC-1521