data: ENOMEM errorpath fixes
If xcalloc() returns NULL, don't try to derrefrence the resulting
pointer in several *_new() "constructors".

Fix a minor memory leak in lmapd_add_config_path() if xcalloc() returns
NULL.

These are considered non-issues, as it only happens when the system is
about to crash/restart due to being out-of-memory.

CEPTRO.br-issue: MDC-1521