lmapd: strict stdin/out/err redirection with error checking
Error-check the redirection of stdin/stdout/stderr to /dev/null when
lmapd is going to run in the background, and abort with an error should
it fail.

Enforce read-only stdin, and write-only stdout/stderr: it helps root out
bugs.

Don't (harmlessly) leak the /dev/null FD (it would be closed later by
the code that ensures only stdin/stdout/stderr are left open in the
background process): it trips static analyzers.

CEPTRO.br-issue: MDC-1521