lmapctl: fix local timezone handling in data.c::set_dateandtime()
The test-suite for the lmap core sets TZ=GMT (it is not even the far
more correct TZ="UTC+00:00"), and thus it is entirely blind to any
timezone issues.

set_dateandtime() in data.c uses mktime(), which ignores the timezone
information in its input and instead applies corrections based on the
local system timezone, which must be undone.  The set_dateandtime() code
was using the wrong sign when applying the correction, but the testsuite
could not notice this.

This is visible in lmapctl, and mentioned as a TODO in the README.md
file: "fix lmapctl timezone bug".

Fortunately, nothing in lmapd itself calls either set_dateandtime() or
one of the data.c functions that do so.  The metadata for reports was
correct all along, and something like "TZ=UTC0 lmapctl report" worked
just fine as a workaround for this bug as far as reports were concerned.

* Fix set_dateandtime() to correct for the local timezone properly, this
  fixes "lmapctl status", and also allows one to use "lmapctl report" to
  generate reports in the local timezone if so desired.

* Detect the presence of timegm() in CMakeLists, and use it instead of
  mktime() where present.

The testsuite will be enhanced to check the base time handling functions
in the lmap core with several timezones in a separate commit.

CEPTRO.br-issue: MDC-1523