tests: use TZ=UTC+00:00 instead of TZ=GMT
TZ=GMT refers to a real place, it can have daylight savings if it
decides to in the future, etc.  Nowadays GMT == UTC+00:00, including the
same leap seconds behaviour.  Still, it is semantically incorrect to ask
for GMT when what you want is actually something else, that just happens
to be what GMT stands for right now.

TZ=UTC+00:00 requests what we actually need as the default local time
for the testsuite.  Use it.

CEPTRO.br-issue: MDC-1523