lmapctl: configurable output width for status subcommand
Add an optional parameter to the -w cli option, for the width of the
output.

When stdout is a tty, use 80 columns as the default when the -w option
is not present, and 132 columns as the default when -w is used without
an argument.

When stdout is not a tty, default to unlimited width.

An argument of zero (-w0) requests unlimited width.  It is not really
unlimited to avoid making the code more complicated: it will be limited
to something a bit smaller than INT_MAX, which should be good enough.

Note that the code doesn't attempt to detect terminal size to use as the
default width when stdout is a terminal: it would need enough CMakeList
and #ifdef gunk so as to not be worth it at this time.